[Backport] 8272873: C2: Inlining should not depend on absolute call site counts
Summary: C2 considers absolute call site counts in its inlining
decisions, which seems very wrong considering the asynchronous nature of
profiling and background compilation (See InlineTree::should_inline()).
It causes substantial over-inlining, which in presence of a depth-first
inlining can lead to an early cut off. It also is inherently unstable.
C2 already uses call frequency as an additional factor and it's better
to consider only that in the inlining heuristic. The author did extensive
benchmarking it yielded almost no losses and single-digit wins (up to
5%) on some benchmarks. The author think it's safe to remove/deprecate
InlineFrequencyCount and continue using InlineFrequencyRatio instead.
The author found that converting the frequency computation to FP and
setting InlineFrequencyRatio=0.25 (inline a method that is called a least
25% of the time) works best.

Test Plan: ci jtreg


Issue: https://github.com/dragonwell-project/dragonwell11/issues/635

+++ src/hotspot/share/runtime/compilationPolicy.hpp
@@ -142,7 +142,7 @@ class StackWalkCompPolicy : public NonTieredCompPolicy {
   // they are used for performance debugging only (print better messages)
   static const char* _msg;            // reason for not inlining
 
-  static const char* shouldInline   (const methodHandle& callee, float frequency, int cnt);
+  static const char* shouldInline   (const methodHandle& callee, double frequency, int cnt);
   // positive filter: should send be inlined?  returns NULL (--> yes) or rejection msg
   static const char* shouldNotInline(const methodHandle& callee);
   // negative filter: should send NOT be inlined?  returns NULL (--> inline) or rejection msg
